properly parse port numbers greater than 32767
Range of valid TCP ports is 1-65535, thus using short as type of port
value prevents user from using full range of valid port numbers.

Port 0 is also valid, but reserved only for requesting a
system-allocated(dynamic) port number from OS, and not valid for use of
host-to-host communication. Thus port 0 shouldan treat it as default
port number of protocol in use when it appears in a URL/URI string.

Signed-off-by: Mehmet Akif Tasova <makiftasova@gmail.com>